requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.10
python-dotenv==1.0.0